test: add interact.test.ts file
