+++ .github/pull_request_template.md
@@ -1,19 +1,27 @@
 ## Type of this PR
+
 > Select at least one type
+
 - [ ] New feature (non-breaking change which adds functionality)
-- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
+- [ ] Breaking change (fix or feature that would cause existing functionality to
+      not work as expected)
 - [ ] Bug fix
-- [ ] Dependency, Environment, Build related update 
+- [ ] Dependency, Environment, Build related update
 - [ ] Documentation
-- [ ] Other tiny fix 
+- [ ] Other tiny fix
+
 ## Describe your changes
+
 ## Issue ticket number and other helpful resource
 
 ## Checklist before requesting a review
+
 - [ ] I have performed a self-review of my code
 - [ ] I have commented my code, particularly in hard-to-understand areas
-- [ ] Leave reference about this pull request, if exist (ex. discord message, google docs, etc)
+- [ ] Leave reference about this pull request, if exist (ex. discord message,
+      google docs, etc)
 - [ ] If it is a core feature, I have added thorough tests.
 
 ## Checklist after creating a pull request
+
 - [ ] Mention reviews in discord channel directly
